requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pytrends==4.9.2
prefect==2.19.9
python-dotenv==1.0.1
